% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_functions.R
\name{agreg_df}
\alias{agreg_df}
\title{aggregate data}
\usage{
agreg_df(df_input, names_vars_by, names_vars_num)
}
\arguments{
\item{df_input}{\code{data.table} data}

\item{names_vars_by}{\code{character} col name to group by}

\item{names_vars_num}{\code{numeric} cols names numerical}
}
\description{
aggregate data
}
